Update camera.py
SUPPORT_STREAM was used from freebox_home, this is a deprecated constant which will be removed in HA Core 2025.1. Use CameraEntityFeature.STREAM instead, please report it to the author of the 'freebox_home' custom integration
SUPPORT_ON_OFF was used from freebox_home, this is a deprecated constant which will be removed in HA Core 2025.1. Use CameraEntityFeature.ON_OFF instead, please report it to the author of the 'freebox_home' custom integration