Update binary_sensor.py
DEVICE_CLASS_MOTION was used from freebox_home, this is a deprecated constant which will be removed in HA Core 2025.1. Use BinarySensorDeviceClass.MOTION instead, please report it to the author of the 'freebox_home' custom integration
DEVICE_CLASS_DOOR was used from freebox_home, this is a deprecated constant which will be removed in HA Core 2025.1. Use BinarySensorDeviceClass.DOOR instead, please report it to the author of the 'freebox_home' custom integration