Update alarm_control_panel.py
STATE_ALARM_PENDING was used from freebox_home, this is a deprecated constant which will be removed in HA Core 2025.11. Use AlarmControlPanelState.PENDING instead, please report it to the author of the 'freebox_home' custom integration
STATE_ALARM_ARMING was used from freebox_home, this is a deprecated constant which will be removed in HA Core 2025.11. Use AlarmControlPanelState.ARMING instead, please report it to the author of the 'freebox_home' custom integration
STATE_ALARM_DISARMING was used from freebox_home, this is a deprecated constant which will be removed in HA Core 2025.11. Use AlarmControlPanelState.DISARMING instead, please report it to the author of the 'freebox_home' custom integration
STATE_ALARM_TRIGGERED was used from freebox_home, this is a deprecated constant which will be removed in HA Core 2025.11. Use AlarmControlPanelState.TRIGGERED instead, please report it to the author of the 'freebox_home' custom integration
DEVICE_CLASS_BATTERY was used from freebox_home, this is a deprecated constant which will be removed in HA Core 2025.1. Use SensorDeviceClass.BATTERY instead, please report it to the author of the 'freebox_home' custom integration